Complete read_firestore console app (#40)
* Setup a Docker container with a console app

I created a simple dart app that reads from the firestore and writes to
a csv file.

I then made a Dockerfile that builds the app and runs it on startup.

* Add retrieveDocument & createDocument functions

* Complete read_firestore console app

We read the preferences data  the firestore then convert it to a matrix
where all unrated items are zero and write the result out to a csv file.